build: setup iOS bridge package
