+++ ios/Bridge/GomobileIPFS/GomobileIPFS.h
@@ -1,0 +1,19 @@
+//
+//  GomobileIPFS.h
+//  Bridge
+//
+//  Created by Antoine Eddi on 2/23/20.
+//  Copyright © 2020 Antoine Eddi. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for GomobileIPFS.
+FOUNDATION_EXPORT double GomobileIPFSVersionNumber;
+
+//! Project version string for GomobileIPFS.
+FOUNDATION_EXPORT const unsigned char GomobileIPFSVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <GomobileIPFS/PublicHeader.h>
+
+
