+++ ios/Bridge/GomobileIPFSTests/GomobileIPFSTests.swift
@@ -1,0 +1,34 @@
+//
+//  GomobileIPFSTests.swift
+//  GomobileIPFSTests
+//
+//  Created by Antoine Eddi on 2/23/20.
+//  Copyright © 2020 Antoine Eddi. All rights reserved.
+//
+
+import XCTest
+@testable import GomobileIPFS
+
+class GomobileIPFSTests: XCTestCase {
+
+    override func setUp() {
+        // Put setup code here.
+    }
+
+    override func tearDown() {
+        // Put teardown code here.
+    }
+
+    func testExample() {
+        // This is an example of a functional test case.
+        XCTAssert(true, "Pass")
+    }
+
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
