--- ios/Example/Podfile
@@ -1,23 +1,0 @@
-use_frameworks!
-
-platform :ios, '10.0'
-
-
-target 'GomobileIPFS_Example' do
-  pod 'GomobileIPFS', :path => '../'
-
-  target 'GomobileIPFS_Tests' do
-    inherit! :search_paths
-
-    pod 'FBSnapshotTestCase' , '~> 2.1.4'
-  end
-
-  post_install do |installer|
-    installer.pods_project.targets.each do |target|
-      target.build_configurations.each do |config|
-        config.build_settings['ENABLE_BITCODE'] = 'NO'
-      end
-    end
-  end
-
-end
