--- ios/README.md
@@ -1,29 +1,0 @@
-# GomobileIPFS
-
-[![CI Status](https://img.shields.io/travis/gfanton/GomobileIPFS.svg?style=flat)](https://travis-ci.org/gfanton/GomobileIPFS)
-[![Version](https://img.shields.io/cocoapods/v/GomobileIPFS.svg?style=flat)](https://cocoapods.org/pods/GomobileIPFS)
-[![License](https://img.shields.io/cocoapods/l/GomobileIPFS.svg?style=flat)](https://cocoapods.org/pods/GomobileIPFS)
-[![Platform](https://img.shields.io/cocoapods/p/GomobileIPFS.svg?style=flat)](https://cocoapods.org/pods/GomobileIPFS)
-
-## Example
-
-To run the example project, clone the repo, and run `pod install` from the Example directory first.
-
-## Requirements
-
-## Installation
-
-GomobileIPFS is available through [CocoaPods](https://cocoapods.org). To install
-it, simply add the following line to your Podfile:
-
-```ruby
-pod 'GomobileIPFS'
-```
-
-## Author
-
-gfanton, guilhem.fanton@gmail.com
-
-## License
-
-GomobileIPFS is available under the MIT license. See the LICENSE file for more info.
